Rename const_static_lifetime to redundant_static_lifetime.
