+++ tests/ui/redundant_static_lifetime.rs
